Only a few files in this modification, but still some quite large changes. Very sorry if this patch messes anything up, I tested for a very long time to make sure nothing could go wrong. If libsamplerate is not found, JTTL Audio should compile fine without it (it had when I tested it)
JTTL's Audio Plugin
- Fixed name output, in some places it was "Sound" before.
- Improved debug output and used Rice Video conventions for printf (the "[plugin name] message" convention)
- Added Libsamplerate support
- Bug fix: Loads configuration before setting frequency (Before the frequency in the config file was not used)
- Added new options, RESAMPLE and TIME_COMPENSATION. Both have options 0-2, see updated jttl_audio.conf for more information.
- Added new resampler (libsamplerate Best Quality SINC resampler), established options for the others (None, and the default, Linear resampling)
- Better GTK Test output.
- Fixed jttl_audio/main.c indentation, because it was indented differently throughout the file. (That will probably make this patch way bigger than it should be. Sorry.)

I will try to do more work on the GUI of JTTL, and maybe add more resampler options. And if anyone can figure out the problem with libresamplerate's resampler causing problems (I'm aware its a memory overflow problem, But I can't quite figure out what to do to fix it, since somehow it manages not to crash anything running, but if I do any of the things that did fix the segfaulting on exit the buffer did not get filled all the way during gameplay. Its an ugly situation and its definitally my fault.) The new resampler is not default, for its problems and its unoptimized nature (I'm not using the streaming APIs of libsamplerate, so it can probably be further optimized.)

Well, Here goes. I'll cross my fingers and hope that nothing was broken in the process of this patch.