More things: - Committed missing files - Added Save/Load/Switch State messages - Added Support for TXT_OBJECT. (Untested!) - More work on the initial support
BTW, This patch does not unify the def of TXT_OBJECT. Why? Because GFX_INFO and the function prototypes are defined twice in the exact same manor - All I can recieve from that is that gfx.h needs to stay different than the original spec file, and so solving the issue of having two defs of TXT_OBJECT is outside the scope of this branch. It's not my decision, but I think it's important to note this anyways.