Wrote message queuing code for OSD. Added support for fade in/out animations. Added more flexible framework for animations so different animations can be added later.
Note: Currently, multiple messages in the same corner of the screen will be rendered over each other instead of moving above or below the other messages. I'll fix this soon. Also, until nmn implements the opengl viewport code in the graphics plugins, the position of the OSD messages will not work correctly.